+++ vtm/src/org/oscim/layers/tile/buildings/BuildingLayer.java
@@ -1,6 +1,6 @@
 /*
  * Copyright 2013 Hannes Janetzek
- * Copyright 2016-2017 devemux86
+ * Copyright 2016-2018 devemux86
  * Copyright 2016 Robin Boldt
  * Copyright 2017 Gustl22
  *
@@ -52,17 +52,15 @@ public class BuildingLayer extends Layer implements TileLoaderThemeHook {
     private static final Object BUILDING_DATA = BuildingLayer.class.getName();
 
     // Can be replaced with Multimap in Java 8
-    protected HashMap<Integer, List<BuildingElement>> mBuildings = new HashMap<>();
+    protected java.util.Map<Integer, List<BuildingElement>> mBuildings = new HashMap<>();
 
     class BuildingElement {
         MapElement element;
         ExtrusionStyle style;
-        boolean isPart;
 
-        BuildingElement(MapElement element, ExtrusionStyle style, boolean isPart) {
+        BuildingElement(MapElement element, ExtrusionStyle style) {
             this.element = element;
             this.style = style;
-            this.isPart = isPart;
         }
     }
 
@@ -103,17 +101,14 @@ public class BuildingLayer extends Layer implements TileLoaderThemeHook {
 
         // Filter all building elements
         // TODO #TagFromTheme: load from theme or decode tags to generalize mapsforge tags
-        boolean isBuildingPart = element.tags.containsKey(Tag.KEY_BUILDING_PART)
-                || (element.tags.containsKey("kind") && element.tags.getValue("kind").equals("building_part")); // Mapzen
-        if (element.tags.containsKey(Tag.KEY_BUILDING) || isBuildingPart
-                || (element.tags.containsKey("kind") && element.tags.getValue("kind").equals("building"))) { // Mapzen
+        if (element.isBuilding() || element.isBuildingPart()) {
             List<BuildingElement> buildingElements = mBuildings.get(tile.hashCode());
             if (buildingElements == null) {
                 buildingElements = new ArrayList<>();
                 mBuildings.put(tile.hashCode(), buildingElements);
             }
             element = new MapElement(element); // Deep copy, because element will be cleared
-            buildingElements.add(new BuildingElement(element, extrusion, isBuildingPart));
+            buildingElements.add(new BuildingElement(element, extrusion));
             return true;
         }
 
@@ -171,7 +166,7 @@ public class BuildingLayer extends Layer implements TileLoaderThemeHook {
         List<BuildingElement> tileBuildings = mBuildings.get(tile.hashCode());
         Set<BuildingElement> rootBuildings = new HashSet<>();
         for (BuildingElement partBuilding : tileBuildings) {
-            if (!partBuilding.isPart)
+            if (!partBuilding.element.isBuildingPart())
                 continue;
 
             String refId = partBuilding.element.tags.getValue(Tag.KEY_REF); // #TagFromTheme
@@ -181,7 +176,7 @@ public class BuildingLayer extends Layer implements TileLoaderThemeHook {
 
             // Search buildings which inherit parts
             for (BuildingElement rootBuilding : tileBuildings) {
-                if (rootBuilding.isPart
+                if (rootBuilding.element.isBuildingPart()
                         || !(refId.equals(rootBuilding.element.tags.getValue(Tag.KEY_ID))))
                     continue;